+++ CHANGELOG.md
@@ -23,7 +23,7 @@ Bugfixes:
   - Fixed ilevel calculation, forgot to double 2handed ilevel as offhand
   - Refactored Paser#get_classes()., Parser#get_gear
 
-## Alpha: 0.0.4 (2015-01-06)
+## Alpha: 0.0.4 (2015-01-07)
 
 Features
   - Added new method to creating a Character class, Character.new("Benpi Kancho"), can pass name without server if you have an extremely rare name
@@ -31,6 +31,7 @@ Features
   - Character, DiscipleList, GearList class have had a to_json method added to them
   - Added first_name and last_name methods
   - Added minion parser + mounts parser
-  - swap to Beacon for testing
 
 Bugfixes
+  - Rewriting more comprehensive tests
+  - Change the way Character.new arguments work